@@ -32,14 +32,14 @@
     "bootstrap": "^4.3.1",
     "core-js": "^3.2.1",
     "file-loader": "^0.11.2",
-    "react": "^16.9.0",
+    "react": "^16.11.0",
     "react-ace": "^7.0.4",
     "react-bootstrap": "^1.0.0-beta.12",
-    "react-dom": "^16.9.0",
+    "react-dom": "^16.11.0",
     "react-redux": "^7.1.1",
     "react-router": "^4.2.0",
     "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.0.1",
+    "react-router-dom": "^5.1.2",
     "redux": "^3.6.0",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
