@@ -46,16 +46,16 @@
     "postcss-loader": "3.0.0",
     "postcss-preset-env": "6.0.6",
     "postcss-safe-parser": "4.0.1",
-    "react": "^16.9.0",
+    "react": "^16.11.0",
     "react-ace": "^7.0.4",
     "react-app-polyfill": "^0.1.3",
-    "react-bootstrap": "^1.0.0-beta.12",
+    "react-bootstrap": "^1.0.0-beta.14",
     "react-dev-utils": "^6.0.4",
-    "react-dom": "^16.9.0",
+    "react-dom": "^16.11.0",
     "react-iframe": "^1.8.0",
     "react-redux": "^7.1.1",
     "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.0.1",
+    "react-router-dom": "^5.1.2",
     "redux": "^3.6.0",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
